requirements.txt
PyJWT==1.7.1
pycryptodome==3.10.4
Faker
colored